projects/aws-rekognition/pyproject.toml
[project]
name = "aws-rekognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.18",
    "boto3-stubs[rekognition]>=1.38.18",
    "pillow>=11.2.1",
]
